Refactor: Rework relationships between the shopping cart entities

+++ src/main/java/com/shopping/cart/validator/ProductValidator.java
@@ -19,7 +19,7 @@ public class ProductValidator {
         if (productRepository.findByName(product.getName()).isPresent()) {
             throw new ProductException("Product with such name is already exists!");
         }
-        validatePrice(product.getPrice());
+        validatePrice(product.getPrice().doubleValue());
         return true;
     }
